Add length to Java CodeArea
