+++ java/src/test/java/com/google/openlocationcode/EncodingTest.java
@@ -25,26 +25,20 @@ public class EncodingTest {
 
   private static class TestData {
 
-    private final String code;
     private final double latitude;
     private final double longitude;
-    private final double decodedLatitudeLo;
-    private final double decodedLatitudeHi;
-    private final double decodedLongitudeLo;
-    private final double decodedLongitudeHi;
+    private final int length;
+    private final String code;
 
     public TestData(String line) {
       String[] parts = line.split(",");
-      if (parts.length != 7) {
+      if (parts.length != 4) {
         throw new IllegalArgumentException("Wrong format of testing data.");
       }
-      this.code = parts[0];
-      this.latitude = Double.valueOf(parts[1]);
-      this.longitude = Double.valueOf(parts[2]);
-      this.decodedLatitudeLo = Double.valueOf(parts[3]);
-      this.decodedLongitudeLo = Double.valueOf(parts[4]);
-      this.decodedLatitudeHi = Double.valueOf(parts[5]);
-      this.decodedLongitudeHi = Double.valueOf(parts[6]);
+      this.latitude = Double.valueOf(parts[0]);
+      this.longitude = Double.valueOf(parts[1]);
+      this.length = Integer.valueOf(parts[2]);
+      this.code = parts[3];
     }
   }
 
@@ -56,7 +50,7 @@ public class EncodingTest {
     BufferedReader reader = new BufferedReader(new InputStreamReader(testDataStream, UTF_8));
     String line;
     while ((line = reader.readLine()) != null) {
-      if (line.startsWith("#")) {
+      if (line.startsWith("#") || line.length() == 0) {
         continue;
       }
       testDataList.add(new TestData(line));
@@ -73,122 +67,20 @@ public class EncodingTest {
     } else {
       testPath = bazelRootPath + "/openlocationcode/test_data";
     }
-    return new File(testPath, "encodingTests.csv");
+    return new File(testPath, "encoding.csv");
   }
 
   @Test
   public void testEncodeFromLatLong() {
     for (TestData testData : testDataList) {
-      int codeLength = testData.code.length() - 1;
-      if (testData.code.contains("0")) {
-        codeLength = testData.code.indexOf("0");
-      }
       Assert.assertEquals(
           String.format(
-              "Latitude %f and longitude %f were wrongly encoded.",
+              "Latitude %f, longitude %f and length %d were wrongly encoded.",
               testData.latitude,
-              testData.longitude),
+              testData.longitude,
+              testData.length),
           testData.code,
-          OpenLocationCode.encode(testData.latitude, testData.longitude, codeLength).toString());
+          OpenLocationCode.encode(testData.latitude, testData.longitude, testData.length).toString());
     }
-  }
-
-  @Test
-  public void testDecode() {
-    for (TestData testData : testDataList) {
-      OpenLocationCode.CodeArea decoded = new OpenLocationCode(testData.code).decode();
-
-      Assert.assertEquals(
-          "Wrong low latitude for code " + testData.code,
-          testData.decodedLatitudeLo,
-          decoded.getSouthLatitude(),
-          PRECISION);
-      Assert.assertEquals(
-          "Wrong high latitude for code " + testData.code,
-          testData.decodedLatitudeHi,
-          decoded.getNorthLatitude(),
-          PRECISION);
-      Assert.assertEquals(
-          "Wrong low longitude for code " + testData.code,
-          testData.decodedLongitudeLo,
-          decoded.getWestLongitude(),
-          PRECISION);
-      Assert.assertEquals(
-          "Wrong high longitude for code " + testData.code,
-          testData.decodedLongitudeHi,
-          decoded.getEastLongitude(),
-          PRECISION);
-    }
-  }
-
-  @Test
-  public void testClipping() {
-    Assert.assertEquals(
-        "Clipping of negative latitude doesn't work.",
-        OpenLocationCode.encode(-90, 5),
-        OpenLocationCode.encode(-91, 5));
-    Assert.assertEquals(
-        "Clipping of positive latitude doesn't work.",
-        OpenLocationCode.encode(90, 5),
-        OpenLocationCode.encode(91, 5));
-    Assert.assertEquals(
-        "Clipping of negative longitude doesn't work.",
-        OpenLocationCode.encode(5, 175),
-        OpenLocationCode.encode(5, -185));
-    Assert.assertEquals(
-        "Clipping of very long negative longitude doesn't work.",
-        OpenLocationCode.encode(5, 175),
-        OpenLocationCode.encode(5, -905));
-    Assert.assertEquals(
-        "Clipping of very long positive longitude doesn't work.",
-        OpenLocationCode.encode(5, -175),
-        OpenLocationCode.encode(5, 905));
-  }
-
-  @Test
-  public void testContains() {
-    for (TestData testData : testDataList) {
-      OpenLocationCode olc = new OpenLocationCode(testData.code);
-      OpenLocationCode.CodeArea decoded = olc.decode();
-      Assert.assertTrue(
-          "Containment relation is broken for the decoded middle point of code " + testData.code,
-          olc.contains(decoded.getCenterLatitude(), decoded.getCenterLongitude()));
-      Assert.assertTrue(
-          "Containment relation is broken for the decoded bottom left corner of code "
-              + testData.code,
-          olc.contains(decoded.getSouthLatitude(), decoded.getWestLongitude()));
-      Assert.assertFalse(
-          "Containment relation is broken for the decoded top right corner of code "
-              + testData.code,
-          olc.contains(decoded.getNorthLatitude(), decoded.getEastLongitude()));
-      Assert.assertFalse(
-          "Containment relation is broken for the decoded bottom right corner of code "
-              + testData.code,
-          olc.contains(decoded.getSouthLatitude(), decoded.getEastLongitude()));
-      Assert.assertFalse(
-          "Containment relation is broken for the decoded top left corner of code " + testData.code,
-          olc.contains(decoded.getNorthLatitude(), decoded.getWestLongitude()));
-    }
-  }
-
-  @Test
-  public void testMaxCodeLength() {
-    // Check that we do not return a code longer than is valid.
-    String code = OpenLocationCode.encode(51.3701125, -10.202665625, 1000000);
-    Assert.assertEquals(
-      "Encoded code should have a length of MAX_DIGIT_COUNT + 1 for the plus symbol",
-      OpenLocationCode.MAX_DIGIT_COUNT + 1,
-      code.length());
-    Assert.assertTrue("Code should be valid.", OpenLocationCode.isValidCode(code));
-    // Extend the code with a valid character and make sure it is still valid.
-    String tooLongCode = code + "W";
-    Assert.assertTrue(
-      "Too long code with all valid characters should be valid.",
-      OpenLocationCode.isValidCode(tooLongCode));
-    // Extend the code with an invalid character and make sure it is invalid.
-    tooLongCode = code + "U";
-    Assert.assertFalse(
-      "Too long code with invalid character should be invalid.",
-      OpenLocationCode.isValidCode(tooLongCode));
   }
 }
